feat: add passing of all task variable via cli arguments
