+++ src/core/markdown-generator.ts
@@ -1,15 +1,7 @@
 import path from 'node:path';
 import Handlebars from 'handlebars';
-import type { FileInfo } from '../types';
+import type { FileInfo, MarkdownOptions } from '../types';
 import { replaceTemplateVariables } from '../utils/template-utils';
-
-export interface MarkdownOptions {
-  template?: string;
-  noCodeblock?: boolean;
-  customData?: Record<string, string>;
-  basePath?: string;
-  lineNumbers?: boolean;
-}
 
 function registerHandlebarsHelpers(
   noCodeblock: boolean,
